Change file names to comply with style guide
