+++ BeagleBone_black/src/Lint.files
@@ -15,11 +15,11 @@ motor_control/motor.cpp
 motor_control/motor_controller.hpp
 motor_control/motor_controller.cpp
 i2c/i2c.hpp
-utils/concurrent/Thread.hpp
-utils/concurrent/Thread.cpp
-utils/concurrent/Lock.hpp
-utils/concurrent/Lock.cpp
-utils/concurrent/ConditionVariable.hpp
-utils/concurrent/ConditionVariable.cpp
-utils/concurrent/Barrier.hpp
-utils/concurrent/Barrier.cpp
+utils/concurrent/thread.hpp
+utils/concurrent/thread.cpp
+utils/concurrent/lock.hpp
+utils/concurrent/lock.cpp
+utils/concurrent/condition_variable.hpp
+utils/concurrent/condition_variable.cpp
+utils/concurrent/barrier.hpp
+utils/concurrent/barrier.cpp
