+++ BeagleBone_black/src/Source.files
@@ -7,10 +7,10 @@ SRCS := \
   state_machine/machine-states.cpp \
   motor_control/motor.cpp \
   motor_control/motor_controller.cpp \
-  utils/concurrent/Thread.cpp \
-  utils/concurrent/Lock.cpp \
-  utils/concurrent/ConditionVariable.cpp \
-  utils/concurrent/Barrier.cpp \
+  utils/concurrent/thread.cpp \
+  utils/concurrent/lock.cpp \
+  utils/concurrent/condition_variable.cpp \
+  utils/concurrent/barrier.cpp \
