+++ BeagleBone_black/src/data/data.hpp
@@ -31,7 +31,7 @@ struct Navigation {
   uint32_t distance;
   uint32_t velocity;
   int32_t acceleration;
-  uint64_t stripe_count;
+  uint32_t stripe_count;
 };
 
 // -----------------------------------------------------------------------------
